Move createHigherOrderComponent to @wordpress/compose
