+++ components/higher-order/with-notices/index.js
@@ -6,7 +6,8 @@ import uuid from 'uuid/v4';
 /**
  * WordPress dependencies
  */
-import { Component, createHigherOrderComponent } from '@wordpress/element';
+import { Component } from '@wordpress/element';
+import { createHigherOrderComponent } from '@wordpress/compose';
 
 /**
  * Internal dependencies
