+++ docs/extensibility/extending-blocks.md
@@ -15,7 +15,7 @@ Used to filter the block settings. It receives the block settings and the name o
 A filter that applies to the result of a block's `save` function. This filter is used to replace or extend the element, for example using `wp.element.cloneElement` to modify the element's props or replace its children, or returning an entirely new element.
 
 #### `blocks.getSaveContent.extraProps`
- 
+
 A filter that applies to all blocks returning a WP Element in the `save` function. This filter is used to add extra props to the root element of the `save` function. For example: to add a className, an id, or any valid prop for this element. It receives the current props of the `save` element, the block type and the block attributes as arguments.
 
 _Example:_
@@ -79,7 +79,7 @@ _Example:_
 ```js
 var el = wp.element.createElement;
 
-var withInspectorControls = wp.element.createHigherOrderComponent( function( BlockEdit ) {
+var withInspectorControls = wp.compose.createHigherOrderComponent( function( BlockEdit ) {
 	return function( props ) {
 		return el(
 			wp.element.Fragment,
@@ -113,7 +113,7 @@ _Example:_
 ```js
 var el = wp.element.createElement;
 
-var withDataAlign = wp.element.createHigherOrderComponent( function( BlockListBlock ) {
+var withDataAlign = wp.compose.createHigherOrderComponent( function( BlockListBlock ) {
 	return function( props ) {
 		var newProps = Object.assign(
 			{},
