+++ editor/components/block-edit/context.js
@@ -6,7 +6,8 @@ import { noop } from 'lodash';
 /**
  * WordPress dependencies
  */
-import { createContext, createHigherOrderComponent } from '@wordpress/element';
+import { createContext } from '@wordpress/element';
+import { createHigherOrderComponent } from '@wordpress/compose';
 
 const { Consumer, Provider } = createContext( {
 	name: '',
