+++ packages/compose/src/if-condition/index.js
@@ -1,7 +1,7 @@
 /**
- * WordPress dependencies
+ * Internal dependencies
  */
-import { createHigherOrderComponent } from '@wordpress/element';
+import createHigherOrderComponent from '../create-higher-order-component';
 
 /**
  * Higher-order component creator, creating a new component which renders if
