+++ packages/compose/src/index.js
@@ -3,6 +3,7 @@
  */
 import { flowRight } from 'lodash';
 
+export { default as createHigherOrderComponent } from './create-higher-order-component';
 export { default as ifCondition } from './if-condition';
 export { default as pure } from './pure';
 export { default as remountOnPropChange } from './remount-on-prop-change';
